misc/HtmlTestSuite/results/test-1337.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>IntersectionObserver reports the exception from its callback in the callback's global object</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
</head><body><iframe srcdoc="&lt;div style=&#34;height: 100px;&#34;&gt;foo&lt;/div&gt;"/>
<iframe/>
<iframe/>
<script>
setup({ allow_uncaught_exception: true });

const onerrorCalls = [];
window.onerror = () =&gt; { onerrorCalls.push("top"); };
frames[0].onerror = () =&gt; { onerrorCalls.push("frame0"); };
frames[1].onerror = () =&gt; { onerrorCalls.push("frame1"); };
frames[2].onerror = () =&gt; { onerrorCalls.push("frame2"); };

async_test(t =&gt; {
  window.onload = t.step_func(() =&gt; {
    const target = frames[0].document.querySelector("div");
    const io = new frames[0].IntersectionObserver(new frames[1].Function(`throw new parent.frames[2].Error("PASS");`));
    io.observe(target);

    t.step_timeout(() =&gt; {
      assert_array_equals(onerrorCalls, ["frame1"]);
      t.done();
    }, 25);
  });
});
</script>
</body></html>
